copy modified Endpoints sample to new location
This change copies the sample from docs/managed_vms/endpoints/ to a new
endpoints/getting-started directory. It also generalizes the sample to
work on all GCP environments by adding a custom Dockerfile for use on
GCE and GKE, as well as creating a separate main function for non-App
Engine environments.