kms: fix casing of "plaintext" and "ciphertext" in variable names
When dealing with encryption, "plaintext" and "ciphertext" are terms of art with
particular meaning. Plaintext is data used as input for encryption, whereas
ciphertext is the output of encryption. Plaintext need not be "plain text" -- it
could be an image, for example.

The example code for other languages are already cased/spelled correctly.