kms: make the samples consistent with the other languages (#294)
In particular, this change:
- changes to use a consistent "ID" suffix for resource name components
- cases "cryptoKey" and "keyRing" consistently with the API

Unlike the other languages, I opted not to move the file I/O into the body of
the encrypt and decrypt functions. I think the file as a whole is structured
idiomatically, so I think the exception is reasonable here. Additionally, the
extra arguments extend the function signature, which will not look great when
rendered in https://cloud.google.com/kms/docs/encrypt-decrypt, which has a
fairly narrow view at under 100 columns.

Tracking bug (internal to Google): http://b/64758639