gettingstarted/bookshelf/app:  more dynamic startup
It appears there's test failures due to connection refused for
TestMainFunc, which invokes the application rather than the other
tests which render handlers.

On the suspicion something else is leaving the static port 8080 in
a temporarily unusuable state, this change picks a port randomly
and then passes it via the PORT environment variable which is used
by appengine.Main to override the serving port.

Also fixes a minor typo in internal/webtest.