spanner: use correct/consistent logic to transfer budget
According to the Cloud Spanner docs [1], the `dmlwritetxn` snippet moves
100000 from Album1 to Album2. However, it was actually moving 200000,
and that only worked if you ran the tests in a specific order.

The `writetransaction` snippet was also moving 200000, but in the opposite
direction, and without confirming that the budget was actually available
before making the transfer.

I cleaned up both snippets, made them consistent with one another, fixed
the logic error in `writetransaction`, and updated the tests to make the
results more predictable.

[1]: https://cloud.google.com/spanner/docs/getting-started/go/