jobs/v4: reorganize tests to improve flakiness.
Previously, these tests leveraged RunMain to construct/deconstruct
resources for each invocation.  Based on the suspicion we're simply
running into indexing delays, I've switched this to a single test file
where resources are recreated with less volatility.

This also fixes the retry for custom attribute ranking, which was
in a retry but didn't leverage the retry semantics so it would always
fail after the first iteration.