dlp: only wait for 1 minute for risk jobs
These have been timing out after 1 hour recently. Switch to only waiting
for one minute for the job to finish and only having the test check the
job was created.

Alternatively, we could pass a context.Context as an argument. But,
users have found that confusing. So, I'm not sure if it's better to do
that or to inline the deadline in the sample. I inlined with an
explanatory comment but am open to other ideas.