fix(bigquery):  minor refactor to avoid pushback failures
Currently, table snippet tests mutate metadata for a table serially.
Recently, changes to how mutation pushback work in the backend are
causing intermittent failures.  The pushback is significant enough
that retries are insufficient to address this with the table snippet
tests.

This change addresses the problem by directing some of the mutation
examples to modify an additional BigQuery table.