test(spanner): fix leaderboard test
The Spanner leaderboard test generates scores for random dates for the past
24 months. The test queried the scores for all records that were not older
than 178 hours, which could return zero records if the random insert process
did not happen to insert a record with a date no older than 178 hours. The
query has now been changed to search for dates that are at least 24 months
ago, to ensure that there will always be results.

Fixes #1820