feat: add h2c sample for Cloud Run
I'm contributing a pure HTTP2 cleartext (h2c) server example as part
of HTTP/2 launch for Cloud Run.

This server doesn't support upgrading from HTTP/1 intentionally to
show that end-to-end HTTP/2 works on Cloud Run.

I intentionally have not added tests as the tests mainly would be
testing x/net/http2 package, which is not relevant to test here,
and would make the code more complicated (to make it testable).

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>