fix: use earliest version time for backups
Using the local date/time as the version time of a backup can cause issues if
there's a difference between the local time and the backend time. The backend
will not accept any version time that is in the future, or that is before the
earliest version time of a database. Using the earliest version time is
therefore the safest option.

Fixes #1900