Add comment about authenticated push JWT token validation.
This snippet is used in the Cloud Pub/Sub docs (https://cloud.google.com/pubsub/docs/push#validating_tokens) and many users are not aware that signature verification of the token is not enough, the claim needs to be validated also.

See similar change in the nodejs repo: GoogleCloudPlatform/nodejs-docs-samples#2222.