test(dataproc): improve flakiness for quickstart
The quickstart is fairly focused on resource deletion (rightly so), but
these opportunistic deletes can cause flakiness.

This change makes cluster deletion in the test non-failing, as well as removing
some unnecessary resource enumeration in the utility cluster deletion
function.  The quickstart itself performs setup and teardown, which
retains the necessary signalling.

Fixes: https://github.com/GoogleCloudPlatform/golang-samples/issues/2199