feat(security-command-center): simplify "create client with endpoint" sample (#4547)
I added this sample in #4419. See that PR for the rationale for creating both regional and global clients. However, when I sent a PR for the Java equivalent of this sample, the reviewer wouldn't accept a sample that created both regional and non-regional clients.

I still think it would be useful to show the creation of both clients, but I think it's more important to make the samples do approximately the same thing in each language.