chore: Update the Spanner dependency to latest release v1.76.1 (#5190)
* Update the Spanner dependency to latest release v1.76.1

V1.76.1 has multiplex session support. So update the samples with latest version so that samples can be executed with Spanner multiplex sessions.

* Update the Spanner dependency to latest release v1.76.1

V1.76.1 has multiplex session support. So update the samples with latest version so that samples can be executed with Spanner multiplex sessions.

* Update the Spanner dependency to latest release v1.76.1

V1.76.1 has multiplex session support. So update the samples with latest version so that samples can be executed with Spanner multiplex sessions.