chore(pubsub): fix error check in bucket creation (#5383)
storage.ErrBucketNotExist is apparently not the top error in the chain when calling Attrs on a bucket, so the error handling path wasn't being taken when a bucket didn't exist. (Discovered while running tests on a new project). Using errors.Is() for comparison instead. Validated this works correctly in a new project, however our usual testing projects won't exercise this code path during testing since the bucket already exists.